use graduated training
The idea behind graduated training is to make the agent learn simple
solves before harder solves. We solve 1 move scrambles, then 2, and
onwards. To "graduate" the agent must solve 100% of validation cases in
18/20 of the past episodes.

In validate, we take max moves as a parameter, and address a false
negative where a zero move count solution was counted as not solved.
This was holding us back since some 2 move scrambles are actually
solved without the agent.

Also in validate, we sample the validation sets so that we don't spend a
tremendous amount of time in validation.